requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
et_xmlfile==2.0.0
fake-useragent==2.2.0
greenlet==3.2.2
idna==3.10
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
playwright==1.52.0
psycopg2-binary==2.9.10
pyee==13.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
